features callback, clustering after adding or removing features, types etc
