+++ source/layers/ClusterLayer.ts
@@ -6,9 +6,9 @@ import {Render} from '../renders/Render';
 import {StaticImageRender} from '../renders/StaticImageRender';
 import {ClusterSymbol} from '../symbols/ClusterSymbol';
 import {FeatureGroup} from '../features/FeatureGroup';
-import {IClusterProvider, GridClusterProvider} from "./clusterProviders/GridClusterProvider";
+import {GridClusterProvider} from "./clusterProviders/GridClusterProvider";
+import {IClusterProvider} from "./clusterProviders/types";
 import {FeaturesAddEvent, FeaturesRemoveEvent} from "./FeatureLayer";
-import {Handler} from "../EventHandler";
 
 export interface ClusterLayerConstructorParams extends LayerConstructorParams {
     clusterSymbol?: Symbol<Feature>;
@@ -21,7 +21,6 @@ export interface ClusterLayerConstructorParams extends LayerConstructorParams {
 export class ClusterLayer extends Layer {
     readonly _clusterSymbol: Symbol<Feature>;
     private _gridClusterProvider: IClusterProvider;
-    private _clusterEventHandlers = [];
     private _clusters: FeatureGroup[] = [];
 
     /**
@@ -54,9 +53,6 @@ export class ClusterLayer extends Layer {
                     cluster.features[0].render(resolution, bbox.crs),
                 );
             } else {
-                this._clusterEventHandlers.forEach(({ type, handler }) =>
-                    !cluster.hasListener(type, handler) && cluster.on(type, handler),
-                );
                 renders = renders.concat(cluster.render(resolution, bbox.crs));
             }
 
@@ -75,17 +71,6 @@ export class ClusterLayer extends Layer {
     getFeatures(bbox: Bbox, resolution: number): FeatureGroup[] {
         if (!this.checkVisibility(resolution)) return [];
         return this._gridClusterProvider.getClusters(bbox, resolution);
-    }
-
-    public addClusterEvent(type: string, handler: Handler): void {
-        this._clusterEventHandlers.push({ type, handler });
-    }
-
-    public removeClusterEvent(type: string, handler: Handler): void {
-        const index = this._clusterEventHandlers.findIndex(item =>  item.type === type && item.handler === handler);
-
-        this._clusterEventHandlers.splice(index, 1);
-        this._clusters.forEach(cluster => cluster.off(type, handler));
     }
 
     /**
